Fix a crash by doing cleanup() before emit.
The ViewResolver sent an emit before doing its internal cleanup, possibly
triggering a crash if the view it was working on had been deleted by the
ViewTab. In the process, sends a copy of the url with its unresolvableUrl signal
instead of a reference, since mUrl is wiped out by cleanup(), which is now
called before the emission of the signal.